refactor: migrate to McpServer register API for tool registration
- Remove manual tool management in Handlers class
- Delete individual tool files and consolidate into direct registration
- Use McpServer.registerTool() and registerResource() methods
- Eliminate manual dispatch logic for tools and resources
- Update dependencies in deno.json and deno.lock
- Improve maintainability and MCP specification compliance
