--- src/tools/index.ts
@@ -1,5 +1,0 @@
-export * from "./getPageTool.ts";
-export * from "./insertLinesTool.ts";
-export * from "./listPagesTool.ts";
-export * from "./searchPagesTool.ts";
-export * from "./types.ts";
